Reverse shell spawns a separate process
